tsio: change the db fields of diff/snapshot to bytearray
On this, we will store zlib compressed json data.

We had the following performances:

class                 diffsize  snapsize           test      time
           TimeSerie   7313964  10970942        bigdata  4.832016
           TimeSerie    413227   2093552  lots_of_diffs  4.997464
   ZlibJsonTimeSerie    930659   1396173        bigdata  4.224456
   ZlibJsonTimeSerie     58710    176341  lots_of_diffs  4.490644
  PlainJsonTimeSerie   7313964  10970942        bigdata  4.623829
  PlainJsonTimeSerie    413227   2093552  lots_of_diffs  4.798412
 ZlibPickleTimeSerie   2388508   3575109        bigdata  4.190854
 ZlibPickleTimeSerie    148847    671794  lots_of_diffs  3.862872
PlainPickleTimeSerie   5048868   7570453        bigdata  3.647770
PlainPickleTimeSerie    376915   1568524  lots_of_diffs  3.488112


So we pick something slightly faster and with a much better storage size
behaviour (best of all). The new TimeSerie is what is listed as
ZlibJsonTimeSerie here.

If pure speed becomes an issue (and portability is not an issue),
the plain pickle strategy looks promising (and it is also
100% lossless).


resolves #23