history: normalize insertion dates from the db to utc
Since straight out of the db we only get an offset,
and the utc timezone is quite implicit. We want it to be explicit
since pandas can change it interpretation from version to version
(e.g. .23 to .24 in a .loc in the presence of a datetime with offset
but no tz).

(we know it is utc because:

 * postgres is configured (see conftest) to UTC
 * we insert as utc
)