tsio/changeset_at: force an order to get the nearest insertion_date
Here the query can be used with inequalities, hence we can get several
answers.

However .scalar always pick one. Unfortunately the lack of explicit
ordering in the query turns this into a implementation-dependant
random pick (postgres tuples are never garanteed to be in the right
order).

Here we accuse .scalar to be a bit complacent ...