storage: reduce the max bucket size
We did a series of measurements with various bucket sizes.

Series profile is a typical meteo forecast: two weeks hourly series
with 4 versions per day. We are supposed to perform well with those.

Useful data per actual stored data (we have some waste):

20:  0.94
50:  0.89
80:  0.84
100:  0.81
150:  0.75
200:  0.69
250:  0.64
500:  0.48
1000:  0.32
2000:  0.2

Obviously the lower the better. From 250 to 150, we get 10 points,
which is good.

Read times:

2000 :  3.7403910830034874
1000 :  3.562026291008806
500 :  3.5152407080022385
250 :  4.0067637500033015
200 :  4.177327500001411
150 :  4.355534500005888
100 :  4.414769833005266
80 :  4.556318332994124
50 :  4.028229417002876
20 :  4.058901540993247

The bigger the better, but the difference is not very significant.

Database size:

50 : 173
80 : 156
100 : 155
150 : 138
250 : 155
350 : 176

When too small, the index grow and the general postgres overhead is
felt. When too big, we suffer from the un-sharing of data (see table
1).

The sweet spot is around 150, so let's pick that !