api/find: return more than just names
We return a small `ts` object carrying:
* the name (mandatory)
* the internal metadata (optional)
* the user metadata (optional)

The two optional fields are controlled by the `meta` flag to the find
api point.

It will be very useful for the web views.

This `ts` structure also could be useful for the .catalog api (but
that would entail a serious change). Or maybe .find will become the
new .catalog (for this to happen we need to provide the series
source).