util/diff and tsio/create: regularize the nan values handling
They were special cased in `diff` and tsio/create, in wrong ways.

There's nothing special about a nan, it's a floating point value.
The erase semantics does not change anything.

At creation time, they seem to be useless. But what if we create a
series whose purpose is to patch another ? The formula system
certainly encourages that kind of thing. So we take the nans and we
don't touch them.