api/update: make erasure update explicit
We introduce a `keepnans` keyword whose default value is False.

By default, series given to .update will be stripped of their Nans,
unless we explicitly tell it it is for point erasure purposes.

It is not only for erasure: nans won't be rejected if they apply to
previously unvalued points.

Why would one want "pure" nans ?  A series can also be used as a patch
for another series (as in the formula `priority` operator) and nans
can make sense in this context (they indeed will erase data).

Because of these changes, in series.sql, tsstart and tsend can be null.