http/server/update: accept passing the series data in the `bseries` field
This won't be limited to severe size restrictions (max_body_size can
be made much larger than the usual buffers for the uri), which make it
easy to hit an http 414 error.

The old path of course remain there.

So we add a new branch within the `json` format branch, to fetch the
series from the `bseries` argument if `series` is unpopulated.

closes #42