http: test that exhibit the current behavior with naive bounds
In the case of tz-aware series, the bounds are intepreted as
tz-aware but in UTC.
We also provide a parmeter "tzone", e.g. 'CET'
Hence a request bounds at midnight will return value at midnight
in UTC and not midnight when localized (here T01:00:00+01:00)

A prefered behavior would be to return the actual bounds of the
series at midnight in the localized referentiel (i.e. T00:00:00+01:00)


related #40