codecs: tweak the system until we have a working model
Most of the assembly work remains to be written - a skeleton of the
logic is provided in the test_read_write_2_versions test.

We now have 3 files, for:
* the revisions (fixed size)
* the tree (fixed size)
* the chunks (variable size)

Any revision points to an unique entry in the tree.

Any tree item points to the start/end position of a chunk in the
chunks, plus a reference to a parent item in the tree (if any).

From this, many things remain to be done. In short order, we need:
* a proper time series deserializer from a list of chunks
* an object to hold together the read/write logic and the codecs

These will lay the ground of the .get and .update logic.

related to #46