tsio/fs: show that our revision search impl. sucks
The base issue is that we only discriminate versions
down to the second. For any smaller deltas, close revisions
will be stamped at the same second.
Also upon inspection it turns out that we walk the revisions
file in the wrong way: we should walk it from most recent
to older.

related to #47