codecs: store the timestamps down to the nanosecond
It's a bit useless as far as real world precision is concerned
but this will help the tests and provides a reliable round-trip option
for the storage, at the cost of more bits.

Hence we solve the issue related to the loss of precision in the previous
commit.

related to #47