storage/fs: try the lz4 and zstandard compressors
The performance implications are very interesting:

Postgres (reference point):
* tsio.timeseries(tsh,othersources=None).update ran in 0.967 seconds.
* tsio.timeseries(tsh,othersources=None).get ran in 0.066 seconds.
* size: 5332

With lz4:

FS1:
* tsio.timeseriesfs1(tsh,othersources=None).update ran in 0.316 seconds.
* tsio.timeseriesfs1(tsh,othersources=None).get ran in 0.031 seconds.
* size: 5835

With zstandard:

FS1:
* tsio.timeseriesfs1(tsh,othersources=None).update ran in 0.359 seconds.
* tsio.timeseriesfs1(tsh,othersources=None).get ran in 0.038 seconds.
* size: 3873

LZ4 and Zstandard are taken as-is, without any attempt to tune them.

Time-wise with lz4 we get .update at 1/3 the time and .get 1/2 the time of
Postgres (lz4), but it consumes 9% more space.

With Zstandard, we get slightly worse performance than with lz4, but
the size aspect is very interesting: it consumes aprox 30% less space.

The current winner is Zstandard, unless with absolutely don't care
about space, which is unclear (I believe we do).

related to #47